Update master.ino
Added the request functionalities into the code, and updated the UI, where when selecting "tinkering corner" or "sandbox", there will be a top screen scrolling down, showing the user the options to either "play", "cancel", "pause", or "resume".

When choosing the "battery" option, there will only have the battery level of the remote control. For now, I am unable to check for battery levels yet, because I don't have a battery on hand at the moment.

When choosing the "open cabinet" option, the remote should open the cabinet mounted on top of the SESTO robot. Further testing is required next week.

Functionalities of the code may be buggy, needs some testing with the SESTO robot next week (probably Tuesday; Wednesday to be the latest to start on testing).

Update on the PCB board:
The PCBWay website has stated that the PCB and the components have already arrived to their facilities, and are now in the process of preparing the production files.

In summary, remote control is mostly done, it is now in the testing phase. Will update ASAP when the remote control UI is completely finish.